--- src/main/java/duke/DukeException.java
@@ -1,12 +1,0 @@
-package duke;
-
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-
-    @Override
-    public String toString() {
-        return "Paiseh... " + this.getMessage();
-    }
-}
