+++ src/main/java/roo/Commands.java
@@ -1,5 +1,8 @@
-package duke;
+package roo;
 
+/**
+ * Represents the commands available in the Roo application.
+ */
 public enum Commands {
     LIST, UNMARK, MARK, DELETE, TODO, DEADLINE, EVENT, END, DATE, CLEAR, UNKNOWN
 }
