+++ src/main/java/roo/Parse.java
@@ -1,11 +1,19 @@
-package duke;
+package roo;
 
-import duke.task.Deadline;
-import duke.task.Event;
-import duke.task.Task;
-import duke.task.Todo;
+import roo.task.Deadline;
+import roo.task.Event;
+import roo.task.Task;
+import roo.task.Todo;
 
+/**
+ * Deals with making sense of the user command.
+ */
 public class Parse {
+    /**
+     * Creates a Task object based on the input string.
+     * @param str The input string representing a task.
+     * @return A Task object parsed from the input string, or null if parsing fails.
+     */
     public static Task makeTask(String str) {
         try {
             if (str.startsWith("[T]")) {
@@ -32,12 +40,17 @@ public class Parse {
                             str.substring(sub2 + 4), false);
                 }
             }
-        } catch (DukeException e) {
+        } catch (RooException e) {
             System.err.println(e.getMessage());
         }
         return null;
     }
 
+    /**
+     * Parses a user input string and determines the corresponding command.
+     * @param input The user input string.
+     * @return The Commands enum representing the detected command.
+     */
     public static Commands parse(String input) {
         Commands c;
         if (input.equals("list")) {
