+++ src/main/java/roo/Storage.java
@@ -1,6 +1,6 @@
-package duke;
+package roo;
 
-import duke.task.Task;
+import roo.task.Task;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -11,15 +11,26 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Scanner;
 
+/**
+ * Deals with loading tasks from the file and saving tasks in the file.
+ */
 public class Storage {
     private File file;
     private String filePath;
 
+    /**
+     * Constructs a Storage object with the specified file path.
+     * @param filePath The path to the file where tasks are stored.
+     */
     public Storage (String filePath) {
         this.filePath = filePath;
         file = new File(this.filePath);
     }
 
+    /**
+     * Initializes the Storage and reads task data from the file if it exists.
+     * @param tasks The TaskList to populate with read tasks.
+     */
     public void initialise(TaskList tasks) {
         try {
             this.file = new File(this.filePath);
@@ -31,6 +42,11 @@ public class Storage {
         }
     }
 
+    /**
+     * Reads the contents of the file and adds tasks to the given TaskList.
+     * @param tasks The TaskList to add the tasks to.
+     * @throws FileNotFoundException If the file is not found.
+     */
     public void readFileContents(TaskList tasks) throws FileNotFoundException {
         Scanner sc = new Scanner(this.file);
         while (sc.hasNext()) {
@@ -42,6 +58,9 @@ public class Storage {
         sc.close();
     }
 
+    /**
+     * Clears the file contents by deleting and recreating the file.
+     */
     public void clear() {
         try {
             Files.delete(Paths.get(filePath));
@@ -51,6 +70,10 @@ public class Storage {
         }
     }
 
+    /**
+     * Modifies the file contents to match the provided list of tasks.
+     * @param tasks The list of tasks to update the file with.
+     */
     public void modifyFile(ArrayList<Task> tasks) {
         try {
             Files.delete(Paths.get(filePath));
@@ -65,6 +88,10 @@ public class Storage {
         }
     }
 
+    /**
+     * Adds a task to the file.
+     * @param task The task to be added to the file.
+     */
     public void add(Task task) {
         try {
             FileWriter writer = new FileWriter(filePath, true);
