+++ src/test/java/roo/ParseTest.java
@@ -1,6 +1,5 @@
-package duke;
+package roo;
 
-import duke.task.Todo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
