+++ src/test/java/roo/task/EventTest.java
@@ -1,6 +1,6 @@
-package duke.task;
+package roo.task;
 
-import duke.DukeException;
+import roo.RooException;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
@@ -16,7 +16,7 @@ public class EventTest {
             Event event = new Event("valid input 1", "30-08-2023 23:59", "31-08-2023 23:59");
             assertEquals("[E][ ] valid input 1 from: 30 Aug 2023 11:59 PM to: 31 Aug 2023 11:59 PM",
                     event.toString());
-        } catch (DukeException exception) {
+        } catch (RooException exception) {
             fail();
         }
     }
@@ -27,7 +27,7 @@ public class EventTest {
             Event event = new Event("valid input 2", "30 Aug 2023 11:59 PM", "31 Aug 2023 11:59 PM");
             assertEquals("[E][ ] valid input 2 from: 30 Aug 2023 11:59 PM to: 31 Aug 2023 11:59 PM",
                     event.toString());
-        } catch (DukeException exception) {
+        } catch (RooException exception) {
             fail();
         }
     }
@@ -37,7 +37,7 @@ public class EventTest {
         try {
             Event event = new Event("", "30-08-2023 23:59", "31-08-2023 23:59");
             fail();
-        } catch (DukeException exception) {
+        } catch (RooException exception) {
             assertEquals("Description is EMPTY!!!\n", exception.getMessage());
         }
     }
@@ -48,7 +48,7 @@ public class EventTest {
             Event event = new Event("valid date 1", "30-08-2023 21:00", "31-08-2023 23:59");
             assertEquals(LocalDateTime.parse("30-08-2023 21:00", DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm")),
                     event.getDate());
-        } catch (DukeException exception) {
+        } catch (RooException exception) {
             fail();
         }
     }
@@ -59,7 +59,7 @@ public class EventTest {
             Event event = new Event("valid date 2", "30 Aug 2023 09:00 PM", "31 Aug 2023 11:59 PM");
             assertEquals(LocalDateTime.parse("30-08-2023 21:00", DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm")),
                     event.getDate());
-        } catch (DukeException exception) {
+        } catch (RooException exception) {
             fail();
         }
     }
